fix: Browser feature compilation errors (Fixes #5)
Root cause: Optional browser dependencies (chromiumoxide, tokio, futures,
async-trait, wasmtime) were placed after WASM32 target section without a
new section header, making them WASM32-only instead of native-only.

Changes:
- Reorganized Cargo.toml to place native-only deps under proper target cfg
- Bumped version to 0.3.1 for crates.io release
- Fixed doctest imports (probar -> jugar_probar)
- Reduced complexity in examples (gui_coverage_report, pixel_coverage_heatmap)
- Implemented JSON loading for coverage CLI (removed TODO)
- Updated pmat-gates.toml with exclusions

+++ crates/probar/examples/pong_simulation.rs
@@ -487,8 +487,8 @@ fn demo_pixel_perfect_pong() {
     // Wilson Score Confidence Interval
     println!("\nWilson Score CI (95%):");
     let ci = ConfidenceInterval::wilson_score(
-        report.covered_cells as u32,
-        report.total_cells as u32,
+        report.covered_cells,
+        report.total_cells,
         0.95,
     );
     println!("  [{:.1}%, {:.1}%]", ci.lower * 100.0, ci.upper * 100.0);
